react for major ui components
