+++ index.html
@@ -19,6 +19,7 @@
       registry.npmjs.cf
       registry.npmjs.org
       sessions.bugsnag.com
+      unpkg.com
       www.google-analytics.com
       www.googletagmanager.com
       www.gravatar.com
@@ -34,109 +35,7 @@
     <link href="./index.css" rel="stylesheet" />
   </head>
 
-  <body class="open">
-    <div id="graph"></div>
-    <div id="progress" style="display:none">
-      <div id="progress_inner"></div>
-    </div>
-
-    <div id="graph-controls" >
-      <button id="zoomWidthButton" title="zoom (fit width)" class="material-icons" style="border-radius: 3px 0 0 3px">swap_horiz</button>
-      <button id="zoomDefaultButton" title="zoom (1:1)" class="material-icons" style="border-width: 1px 0px; border-radius: 0">search</button>
-      <button id="zoomHeightButton" title="zoom (fit height)" class="material-icons" style="border-radius: 0 3px 3px 0">swap_vert</button>
-      <button id="downloadButton" title="download as SVG" class="material-icons" style="margin-left: 0.5em">cloud_download</button>
-    </div>
-
-    <div id="inspector">
-      <div id="tabs">
-        <div class="arrow" id="toggleInspectorButton">&#x25c0</div>
-        <div class="button" data-pane="pane-module">Module</div>
-        <div class="button" data-pane="pane-graph">Graph</div>
-        <div class="button active" data-pane="pane-info">&#9432;</div>
-        <input type="text" id="searchText" placeholder="&#x1F50D; &nbsp;Enter module name" />
-      </div>
-
-      <div class="pane" id="pane-graph">
-        <section>
-          <h2>Tools<span class="material-icons"></span></h2>
-          <div>
-            Include:
-            <style>
-              .depInclude  {
-                padding-left: 1em;
-                font-family: monospace;
-              }
-            </style>
-            <label class="depInclude"><input type="checkbox" data-type="dependencies" checked /><code>dependencies</code></label>
-            <label class="depInclude"><input type="checkbox" data-type="devDependencies" /><code>devDependencies</code> (<span style="color: red">&xrarr;</span>)</label>
-            <label class="depInclude"><input type="checkbox" data-type="peerDependencies" /><code>peerDependencies</code> (<span style="color: green">&xrarr;</span>)</label>
-            <!-- It looks like NPM omits these, even if provided by module authors? 
-              <label class="depInclude"><input type="checkbox" data-type="optionalDependencies" /><code>optionalDependencies</code> (<span>- - &rarr;</span>)</label>
-              <label class="depInclude"><input type="checkbox" data-type="optionalDevDependencies" /><code>optionalDevDependencies</code>(<span style="color: red">- - &rarr;</span>)</label>
-            -->
-          </div>
-          <label style="display:block"><input type="checkbox" id="colorize" />Colorize by <a href="https://github.com/npms-io/npms-analyzer" target="_blank">npms.io score</a></label>
-          <label style="display:block"><input type="checkbox" id="busFactor" />Show modules with 0-1 maintainer</label>
-        </section>
-        <section class="dependencies">
-          <h2>Dependencies<span class="material-icons"></span></h2>
-        </section>
-        <section class="maintainers">
-          <h2>Maintainers<span class="material-icons"></span></h2>
-        </section>
-        <section>
-          <h2>Licenses<span class="material-icons"></span></h2>
-          <div class="licenses"></div>
-          <div id="chart">
-          </div>
-        </section>
-      </div>
-
-      <div class="pane" id="pane-module">
-        <h2>Module</h2>
-        <section class="description">
-          <h2>Description<span class="material-icons"></span></h2>
-        </section>
-        <section class="stats">
-          <h2>Stats<span class="material-icons"></span></h2>
-        </section>
-        <section>
-          <h2>package.json<span class="material-icons"></span></h2>
-          <style>
-            .json {font-family: monospace}
-            .json-object, .json-array {
-              list-style: none outside;
-              margin: 0;
-              padding: 0 0 0 1em;
-            }
-            .json-key {color: green;}
-            .json-string {color: blue}
-            .json-boolean {color: purple}
-            .json-number {color:orange}
-            .json-null {color: gray}
-          </style>
-          <div class="json"></div>
-        </section>
-      </div>
-
-      <div class="pane open" id="pane-info">
-        <p>
-        Enter NPM module name here <i class="material-icons">arrow_upward</i> to see the dependency graph.  To graph multiple modules separate names with commas (e.g. <a href="?q=mocha, chalk, rimraf">"mocha, chalk, rimraf"</a>).
-        </p>
-        <div id="drop_target" style="text-align: center">
-					... or drop a <code>package.json</code> file here
-				</div>
-        <div id="info-footer">
-          <div id="copyright">
-            &copy; Robert Kieffer, 2020  MIT License
-            <a id="github" target="_blank" href="https://github.com/broofa/npmgraph">GitHub</a>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div id="errlog" />
-  </body>
+  <body class="open"></body>
 </html>
