+++ js/Store.js
@@ -1,12 +1,11 @@
 import { ajax, report } from './util.js';
 import Module from './Module.js';
 import Flash from './Flash.js';
-import * as semver from '/vendor/semver.js';
+import * as semver from '../vendor/semver.js';
 
-window.semver = semver;
-
-const _getCache = {};
+const _requestCache = {};
 const stats = { active: 0, complete: 0 };
+
 // fetch module url, caching results (in memory for the time being)
 function fetchModule(name, version) {
   const isScoped = name.startsWith('@');
@@ -17,7 +16,7 @@ function fetchModule(name, version) {
   const pathAndVersion = `${path}/${version}`;
 
   // Use cached request if available.  (We can get module info from versioned or unversioned API requests)
-  let req = _getCache[pathAndVersion] || _getCache[path];
+  let req = _requestCache[pathAndVersion] || _requestCache[path];
 
   if (!req) {
     // If semver isn't valid (i.e. not a simple, canonical version - e.g.
@@ -26,16 +25,16 @@ function fetchModule(name, version) {
     // Also, we can't fetch scoped modules at specific versions.  See https://goo.gl/dSMitm
     const reqPath = !isScoped && versionIsValid ? pathAndVersion : path;
 
-    req = _getCache[reqPath] = ajax('GET', `https://registry.npmjs.cf/${reqPath}`);
+    req = _requestCache[reqPath] = ajax('GET', `https://registry.npmjs.cf/${reqPath}`);
 
     req.finally(() => {
       stats.active--;
       stats.complete++;
-      Store.onRequest(stats);
+      Store.onRequest?.(stats);
     });
 
     stats.active++;
-    Store.onRequest(stats);
+    Store.onRequest?.(stats);
   }
 
   return req.then(body => {
@@ -72,41 +71,27 @@ function fetchModule(name, version) {
   });
 }
 
-/**
- * HTTP request api backed by localStorage cache
- */
-export default class Store {
-  static init() {
-    this._inflight = {};
-    this._moduleCache = {};
-    this._requestCache = {};
-    this._noCache = /noCache/i.test(location.search);
-  }
-
-  // GET package info
-  static async getModule(name, version) {
+const _moduleCache = {};
+const Store = {
+  async getModule(name, version) {
     const cacheKey = `${name}@${version}`;
 
-    if (this._moduleCache[cacheKey]) return this._moduleCache[cacheKey];
-
-    let moduleInfo;
-
-    if (!(/github:|git\+/.test(version))) { // We don't support git-hosted modules
-      try {
-        // If semver isn't valid (i.e. not a simple, canonical version - e.g.
-        // "1.2.3") fetch all versions (we'll figure out the specific version below)
-        //
-        // Also, we can't fetch scoped modules at specific versions.  See https://goo.gl/dSMitm
-        moduleInfo = await fetchModule(name, version);
-      } catch (err) {
-        if ('status' in err) {
-          Flash(err.message);
-        } else {
-          report.error(err);
-        }
-      }
+    if (!_moduleCache[cacheKey]) {
+      _moduleCache[cacheKey] = fetchModule(name, version)
+        .then(moduleInfo => {
+          return new Module(moduleInfo);
+        })
+        .catch(err => {
+          if ('status' in err) {
+            Flash(err.message);
+          } else {
+            report.error(err);
+          }
+        });
     }
 
-    return this._moduleCache[cacheKey] = new Module(moduleInfo);
+    return _moduleCache[cacheKey];
   }
-}
+};
+
+export default Store;
